Move `join_bids_path()` helper into `entities`
