+++ bids2table/helpers.py
@@ -1,33 +1,4 @@
-from pathlib import Path
-from typing import Any, Dict, Optional, Union
-
 import pandas as pd
-
-from bids2table.entities import BIDSEntities
-
-
-def join_bids_path(
-    row: Union[pd.Series, Dict[str, Any]],
-    prefix: Optional[Union[str, Path]] = None,
-    valid_only: bool = True,
-) -> Path:
-    """
-    Reconstruct a BIDS path from a table row/record or entities dict.
-
-    Example::
-
-        df = pd.read_parquet("dataset.parquet")
-        paths = df.apply(join_bids_path, axis=1)
-    """
-    if "entities" in row:
-        row = row["entities"]
-
-    if isinstance(row, pd.Series):
-        row = row.to_dict()
-
-    entities = BIDSEntities.from_dict(row, valid_only=valid_only)
-    path = entities.to_path(prefix=prefix, valid_only=valid_only)
-    return path
 
 
 def flat_to_multi_columns(df: pd.DataFrame, sep: str = "__") -> pd.DataFrame:
